Merkle Revert (13): Revert #512 FileBackend for Merkle{Reader,Writer}
Revert #512 as part of backing out our first Merkle tree redesign attempt (13 of 15).
We're planning a simpler approach for the next attempt.

Things that we didn't revert from the original PR:
- EMerkleTreeNode::from_type_and_bytes, which builds a node from its on-disk type byte
  and payload. Why: it's a backend-independent decoding helper on the node enum, useful
  anywhere raw node bytes are turned back into a node.
- The split of MerkleNodeLookup's file read from its byte deserialization — a load() that
  reads the file and a deserialize(Vec<u8>) that decodes the bytes. Why: separating I/O
  from pure decoding lets callers deserialize already-in-memory bytes and makes the
  decoder unit-testable without touching the filesystem.
- #[inline(always)] on the per-node deserialize methods. Why: these are tiny decoders run
  once per node on every tree read, so inlining them trims call overhead on the hot
  traversal path.

ENG-1143
