+++ crates/liboxen/src/core/db/merkle_node.rs
@@ -1,4 +1,3 @@
-pub mod file_backend;
 pub mod merkle_node_db;
 
 pub(crate) use merkle_node_db::MerkleNodeDB;
